requirements.txt
streamlit
pandas
plotly
pymongo
python-dotenv
requests
beautifulsoup4
schedule
